Converted client entity action packet
